Period Report
Addition of period report to the fitting summary. This allows to see
that is the resulting period after the fitting I still need to add at
what point the period was measured. Another things that I have to add
is the dampening cost such that I can avoid using the amplitude
constrain when the network is in constant conditions.